+++ src/Helper/Charset.php
@@ -301,7 +301,7 @@ class Charset
                     }
                     if ($data === false) {
                         $escapedData = '';
-                        $this->getLogger()->errorLog('XML-RPC: ' . __METHOD__ . ": Converting from $srcEncoding to $destEncoding via mbstring: failed...");
+                        $this->getLogger()->error('XML-RPC: ' . __METHOD__ . ": Converting from $srcEncoding to $destEncoding via mbstring: failed...");
                     } else {
                         if ($srcEncoding === 'UTF-8') {
                             $escapedData = $data;
@@ -311,7 +311,7 @@ class Charset
                     }
                 } else {
                     $escapedData = '';
-                    $this->getLogger()->errorLog('XML-RPC: ' . __METHOD__ . ": Converting from $srcEncoding to $destEncoding: not supported...");
+                    $this->getLogger()->error('XML-RPC: ' . __METHOD__ . ": Converting from $srcEncoding to $destEncoding: not supported...");
                 }
         }